Forms: ASCVD Risk Calculators
Add a Form to calculate several 10 year ASCVD Risk scores:

1. ACC/AHA 2013 ASCVD Risk Score
2. Framingham 2008 ASCVD Risk Score (with labs)
3. Framingham 2008 ASCVD Risk SCore (with no labs, uses BMI)
4. Framingham 1998 Hard Coronary Disease Risk Score
5. MESA 2015 CVD Risk Score
6. MESA 2015 CVD with CAC Risk Score

All risk scores are calculated simultaneously in a single form
as long as the required parameters are satisfied. If the parameters
are not satisfied, the form will display the missing parameters or
the range of the parameters.

The files where initially created by formscript.pl, found under
OpenEMR's contrib directory.

The calculators' algorithms and coefficients are available publicly,
and the source literature are included in the save.php file, above
each function. Most of the algorithms are converted to PHP from the
R language and are patterned on the package CVRisk by
Victor Castro <vcastro at mgh.harvard.edu> available at
https://cran.r-project.org/package=CVrisk. This package is also
licensed under GPL-3.

To enable the form, copy the directory openemr/contrib/forms/ASCVD_Risk_Calculators
to openerm/interface/forms. Register it in Admin->Forms Administration, install the
database, and enable the form.